pyproject.toml
[tool.poetry]
name = "backend-template"
version = "0.1.0"
description = ""
authors = ["Serkan Kalay <serkanosmankalay@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12.1,<4.0"
importlib-metadata = {version = ">=1,<5", python = "<3.8"}
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
psycopg2-binary = "^2.9.9"
sqlalchemy = { extras = ["mypy"], version = "^2.0.29" }
pyintervals = "^1.0.1"
python-dateutil = "^2.9.0.post0"
types-python-dateutil = "^2.9.0.20240316"
python-multipart = "^0.0.9"
caldav = "^1.3.9"
icalendar = "^5.0.13"
pydantic = "^2.9.0"
pydantic-settings = "^2.4.0"
celery = "^5.4.0"
redis = "^5.2.0"
filelock = "^3.16.1"
pyyaml = "^6.0.2"
cachetools = "^5.5.0"
pytest-mock = "^3.14.0"
jwt = "^1.3.1"
passlib = "^1.7.4"
pyjwt = "^2.10.1"
bcrypt = "^4.2.1"
typeguard = "^4.4.1"
python-dotenv = "^1.1.1"
celery-stubs = "^0.1.3"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.4"
pytest-cov = "^4.1.0"
pytest-benchmark = "^4.0.0"

[tool.poetry.group.typecheck.dependencies]
mypy = "^1.9"

[tool.poetry.group.linting.dependencies]
black = ">=23.12,<24.0"
flake8 = ">=7,<8.0.0"
isort = "^5.13.2"

[tool.poetry.group.dev.dependencies]
tox = ">=4.12,<5.0"
alembic = ">=1.13.1,<2.0"
import-linter = "^2.0"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
\.eggs
| \.git
| \.mypy_cache
| \.tox
| \.venv
| _build
| build
| dist
)/
'''

[tool.isort]
line_length = 79
profile = 'black'
add_imports = ['from __future__ import annotations']

[tool.mypy]
disable_error_code = "import-untyped"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
